Refactor watcher operators so the base definitions can be reused by the K8s watcher operator
